refactor: client 패키지 이동
