+++ src/main/java/kitchenpos/infra/client/DefaultKitchenridersClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra;
+package kitchenpos.infra.client;
 
 import org.springframework.stereotype.Component;
