+++ src/main/java/kitchenpos/infra/client/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra;
+package kitchenpos.infra.client;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
